Docker-Configuration/requirements.txt
confluent-kafka
kafka-python
pandas
numpy
pyspark
requests
pyodbc
lifetimes